moved login/logout theme change from session code
